New algorithm for averaging patch data in averageAlongAxis.py using cut-planes
Algorithm choice controlled with patchalg keyword in the config
"line" for the old algorithms using line sampling. "cut" for the new using cut-planes.
The line algorithm can fail if the patch cell centers are not perfectly aligned along a line.
The cut algorithm may fail if the boundary is curved in a peculiar way making the cutting plane
cut it in two places.
The cut algorith is slower but considreed more robust, it is now the default.